Extract base controller for Audit controllers
We have common behaviour for:

- Layout
- Helper methods for parameters
